feat: add bulk delete, remove bulk delete params and modify delete_messages_by_author
